+++ CHANGES.md
@@ -14,6 +14,7 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
   such as `action=query`’s `prop`, `list` and `meta` parameters.
   For more information, see the updated README,
   especially the new “automatically combining requests” section.
+- Updated dependencies.
 
 ## v0.2.1 (2021-10-09)